requirements.txt
streamlit
pandas
Pillow
tensorflow-cpu
numpy